refactor: rename Blocks to BlockList
