+++ main.tsx
@@ -2,12 +2,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 
-import blocks from "./dist/components/Blocks/Blocks.json";
+import blocks from "./dist/components/Blocks/BlockList.json";
 
-import { Blocks } from "./dist/components/Blocks/Blocks";
+import { BlockList } from "./dist/components/Blocks/BlockList";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <Blocks blocks={blocks} />
+    <BlockList blocks={blocks} />
   </React.StrictMode>
 );
