+++ src/components/Blocks/BlockList.stories.tsx
@@ -1,5 +1,5 @@
-import { Blocks } from "./Blocks.js";
-import BlocksObject from "./Blocks.json";
+import { BlockList } from "./BlockList.js";
+import BlocksObject from "./BlockList.json";
 import { OpenedHeading1 } from "./Custom/OpenedHeading1.js";
 import { OpenedHeading2 } from "./Custom/OpenedHeading2.js";
 import { OpenedHeading3 } from "./Custom/OpenedHeading3.js";
@@ -8,15 +8,15 @@ import { OpenedToggle } from "./Custom/OpenedToggle.js";
 import type { Meta, StoryObj } from "@storybook/react";
 import type { BlockObject } from "@udus/notion-libs";
 
-const meta: Meta<typeof Blocks> = {
+const meta: Meta<typeof BlockList> = {
   title: "Blocks",
-  component: Blocks,
+  component: BlockList,
   tags: ["autodocs"],
 };
 
 export default meta;
 
-type Story = StoryObj<typeof Blocks>;
+type Story = StoryObj<typeof BlockList>;
 
 export const Default: Story = {
   args: {
