+++ src/components/Blocks/Custom/OpenedHeading1.tsx
@@ -1,5 +1,5 @@
 import { RichText } from "../../RichText/RichText.js";
-import { Blocks } from "../Blocks.js";
+import { BlockList } from "../BlockList.js";
 
 import type { BlockProps } from "../Block.js";
 import type { Heading1BlockObject } from "@udus/notion-libs";
@@ -22,7 +22,7 @@ export const OpenedHeading1: React.FC<Props> = ({ block }) => {
           </summary>
           <div className="notion_toggle_details">
             {block.heading_1.children && (
-              <Blocks blocks={block.heading_1.children} />
+              <BlockList blocks={block.heading_1.children} />
             )}
           </div>
         </details>
