+++ src/components/Blocks/NumberedListItem.tsx
@@ -1,6 +1,6 @@
 import { RichText } from "../RichText/RichText.js";
 
-import { Blocks } from "./Blocks.js";
+import { BlockList } from "./BlockList.js";
 
 import type { BlockProps } from "./Block.js";
 import type { NumberedListItemBlockObject } from "@udus/notion-libs";
@@ -16,7 +16,7 @@ export const NumberedListItem: React.FC<Props> = ({ block }) => {
       <RichText richText={block.numbered_list_item.rich_text} />
       <div>
         {block.numbered_list_item.children && (
-          <Blocks blocks={block.numbered_list_item.children} />
+          <BlockList blocks={block.numbered_list_item.children} />
         )}
       </div>
     </li>
