+++ src/components/Blocks/SyncedBlock.tsx
@@ -1,5 +1,5 @@
 import { Block, type BlockProps } from "./Block.js";
-import { Blocks } from "./Blocks.js";
+import { BlockList } from "./BlockList.js";
 
 import type { SyncedBlockBlockObject } from "@udus/notion-libs";
 
@@ -11,7 +11,7 @@ export const SyncedBlock: React.FC<Props> = ({ block, blocks }) => {
       {block.synced_block.synced_from === null
         ? block.synced_block.children && (
             <div className={"notion_synced_block_original"}>
-              <Blocks blocks={block.synced_block.children} />
+              <BlockList blocks={block.synced_block.children} />
             </div>
           )
         : block.synced_block.synced_from.block && (
