+++ src/components/Blocks/index.ts
@@ -1,6 +1,6 @@
 export * from "./Audio.js";
 export * from "./Block.js";
-export * from "./Blocks.js";
+export * from "./BlockList.js";
 export * from "./Bookmark.js";
 export * from "./Breadcrumb.js";
 export * from "./BulletedList.js";
