+++ src/components/index.tsx
@@ -1,3 +1,3 @@
-export * from "./Blocks/Blocks.js";
+export * from "./Blocks/BlockList.js";
 export * from "./RichText/RichText.js";
 export * from "./Link.js";
